requirements.txt
flask
flask-cors
tensorflow==2.17.0
opencv-python-headless
pillow
numpy
gunicorn 